requirements.txt
# Core
pytz==2018.5
Django==2.1.1
django-configurations==2.1
gunicorn==19.8.1
newrelic==4.4.0.103
django-cors-headers==2.4.0 
djoser==1.1.5

# For the persistence stores
psycopg2-binary==2.7.5
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==2.0.0

# Developer Tools
ipdb==0.11
ipython==6.5.0
mkdocs==0.17.5
flake8==3.5.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1
freezegun==0.3.10

# Static and Media Storage
django-storages==1.7.1
boto3==1.9.2
